# pkg/perf: Optimize perf map parsing