Avoid allocations and remove unnecessarily generic functions for parsing memory addresses.

```
benchstat old.txt new.txt
name             old time/op  new time/op  delta
PerfMapParse-10  5.62µs ± 1%  3.86µs ± 1%  -31.21%  (p=0.000 n=10+10)
```

I think there is still quite a bit more improvement possible, like using a btree instead of a regular slice (since we want it to be sorted afterward), and reusing/pooling buffered readers.